feat: add designation field to user profile
This commit adds a new field `designation` to the user profile.

The `designation` field represents the job title of the user. It is added to the `users` table and is included in the user profile resource.

This change allows users to specify their job title, which can be useful for displaying more detailed user information.