feat(posts): add image upload to cloudinary
This commit introduces the ability for users to upload images to Cloudinary when creating a new post.

The following changes were made:

- Added a new `gif_url` field to the `PostsPostRequest` class to allow users to specify a URL for a GIF attachment.
- Updated the `store` method in the `PostsController` class to handle image uploads using Cloudinary's Laravel SDK.
- Removed the `posted_by` field from the `PostsPostRequest` class as it is now automatically set to the authenticated user's ID.
- Updated the `store` method to deduct the correct number of points from the user's account when creating a post.
- Updated the `store` method to save the image to Cloudinary and store the secure URL in the database.

These changes improve the user experience by allowing users to easily add images to their posts. It also improves the performance and scalability of the application by offloading image storage and delivery to Cloudinary.