feat: implement scheduled credits reset
This commit implements a scheduled command that resets user vote credits back to the default value at the end of each month.

The `ResetCreditsCommand` is scheduled to run on the last day of the month at 11:50 PM. It iterates through all user points records and sets the credits to the default value specified in the command.